fix phone field for postgres

+++ db/migrate/20140727230514_add_new_fields_to_users.rb
@@ -2,6 +2,6 @@ class AddNewFieldsToUsers < ActiveRecord::Migration
   def change
     add_column :users, :alt_email, :string
     add_column :users, :zip, :integer
-    add_column :users, :phone, :integer
+    add_column :users, :phone, :bigint
   end
 end
